+++ flixopt/elements.py
@@ -4,12 +4,12 @@ This module contains the basic elements of the flixopt framework.
 
 from __future__ import annotations
 
-import logging
 import warnings
 from typing import TYPE_CHECKING
 
 import numpy as np
 import xarray as xr
+from loguru import logger
 
 from . import io as fx_io
 from .config import CONFIG
@@ -35,8 +35,6 @@ if TYPE_CHECKING:
         Numeric_TPS,
         Scalar,
     )
-
-logger = logging.getLogger('flixopt')
 
 
 @register_class_for_io
@@ -532,8 +530,8 @@ class Flow(Element):
         if np.any(self.relative_minimum > 0) and self.on_off_parameters is None:
             logger.warning(
                 f'Flow {self.label_full} has a relative_minimum of {self.relative_minimum} and no on_off_parameters. '
-                f'This prevents the flow_rate from switching off (flow_rate = 0). '
-                f'Consider using on_off_parameters to allow the flow to be switched on and off.'
+                f'This prevents the Flow from switching off (flow_rate = 0). '
+                f'Consider using on_off_parameters to allow the Flow to be switched on and off.'
             )
 
         if self.previous_flow_rate is not None: